Adds the basic support for mobile wallet
- Create account (wallet) that is compatible with full node wallets. Use same mechanism for encrypted wallet.
- Persist the extpubkey and wallet secret on IndexedDB/localStorage.
- Allow login and unlock of wallets.
- Generate address locally based on the extpubkey.

This is far from a complete mobile wallet implementation, but it is the beginning of it. Needs to perform blockchain search against an indexer to find used addresses, and needs valid UTXOs to perform transaction generation. Most likely it will ask the full node to make the tx with the specified UTXO, and then receives the tx for signing which must happen on the client.